update: Configured my arch linux box and cleaned some configs
I replacing my entire dotfiles with this. I have provisioned a arch
linux box. I've gotten tired of setting up the same dependecies on each
machine, from a macbook, macstudio, beelink, raspberrypi, etc. This
should be able to consolidate. However the install script needs to be
modified to support homebrew for macOS and raspbian/ubuntu. We want to
be able to clone, install, and stow. Fire and Forget.